+++ common/src/main/java/net/berserker_rpg/effect/BloodReckoningEffect.java
@@ -1,6 +1,7 @@
 package net.berserker_rpg.effect;
 
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 
@@ -11,8 +12,9 @@ public class BloodReckoningEffect extends StatusEffect {
         super(category, color);
     }
 
-    public void onApplied(LivingEntity entity, int amplifier) {
-        super.onApplied(entity, amplifier);
+    @Override
+    public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
+        super.onApplied(entity, attributes, amplifier);
         float missing_health = entity.getMaxHealth() - entity.getHealth();
         float absorption = missing_health * tweaksConfig.value.blood_reckoning_missing_health_to_absorption;
         entity.setAbsorptionAmount(Math.max(entity.getAbsorptionAmount(), absorption));