+++ common/src/main/java/net/berserker_rpg/effect/OutrageEffect.java
@@ -1,6 +1,7 @@
 package net.berserker_rpg.effect;
 
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.entity.effect.StatusEffectInstance;
@@ -20,17 +21,17 @@ public class OutrageEffect extends StatusEffect {
     }
 
     @Override
-    public void onApplied(LivingEntity entity, int amplifier) {
-        super.onApplied(entity, amplifier);
+    public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
+        super.onApplied(entity, attributes, amplifier);
         if (entity.getWorld().isClient()) return;
         if (APPLYING_RAGE_BONUS.remove(entity.getUuid())) return;
 
-        if (entity.hasStatusEffect(BerserkerEffects.RAGE.entry)) {
-            var current = entity.getStatusEffect(BerserkerEffects.OUTRAGE.entry);
+        if (entity.hasStatusEffect(BerserkerEffects.RAGE.effect)) {
+            var current = entity.getStatusEffect(BerserkerEffects.OUTRAGE.effect);
             if (current != null) {
                 int bonusTicks = Math.round(tweaksConfig.value.outrage_rage_bonus_duration * 20F);
                 APPLYING_RAGE_BONUS.add(entity.getUuid());
-                entity.addStatusEffect(new StatusEffectInstance(BerserkerEffects.OUTRAGE.entry, current.getDuration() + bonusTicks, amplifier));
+                entity.addStatusEffect(new StatusEffectInstance(BerserkerEffects.OUTRAGE.effect, current.getDuration() + bonusTicks, amplifier));
             }
         }
     }