+++ common/src/main/java/net/berserker_rpg/item/armor/NorthlingArmor.java
@@ -1,11 +1,10 @@
 package net.berserker_rpg.item.armor;
 
 import net.minecraft.item.ArmorMaterial;
-import net.minecraft.registry.entry.RegistryEntry;
 import net.spell_engine.rpg_series.item.Armor;
 
 public class NorthlingArmor extends Armor.CustomItem {
-    public NorthlingArmor(RegistryEntry<ArmorMaterial> material, Type slot, Settings settings) {
+    public NorthlingArmor(ArmorMaterial material, Type slot, Settings settings) {
         super(material, slot, settings);
     }
 }